simplified files and simple refactoring
