+++ lib/balancer.js
@@ -1,7 +1,10 @@
-const request = require('./request.js');
 const Bottleneck = require('bottleneck');
 const ProxyAgent = require('simple-proxy-agent');
-const utils = require('./utils.js');
+const fetch = require('node-fetch');
+
+function delay(ms) {
+  return new Promise(resolve => setTimeout(resolve, ms));
+}
 
 const defaultOptions = {
   poolExpired: 1 * 60 * 1000,
@@ -23,18 +26,19 @@ class Balancer {
       minTime: options.minTime || defaultOptions.minTime
     });
     this.fetching = false;
+    this.requestor = fetch
   }
 
   async getProxies() {
-    if(this.fetching) {
-      if(this.proxies.length > 0) {
+    if (this.fetching) {
+      if (this.proxies.length > 0) {
         return this.proxies;
       }
-      await utils.delay(200);
+      await delay(200);
       return this.getProxies();
     }
 
-    if(!this.lastUpdate || this.proxies.length === 0 || Date.now() > this.lastUpdate + this.options.poolExpired) {
+    if (!this.lastUpdate || this.proxies.length === 0 || Date.now() > this.lastUpdate + this.options.poolExpired) {
       this.fetching = true;
       try {
         const proxies = await this.options.proxyFn();
@@ -58,7 +62,7 @@ class Balancer {
         timeout: this.options.proxyTimeout
       });
       const res = await this.limiter.schedule(() => {
-        return request(url, {
+        return fetch(url, {
           agent: agent,
           ...options
         }, timeout)
@@ -66,7 +70,7 @@ class Balancer {
       return res;
     } catch (err) {
       // Retry if proxy error
-      if(err.name && (err.name === "FetchError" || err.name === "AbortError")) {
+      if (err.name && (err.name === "FetchError" || err.name === "AbortError")) {
         return this.request(url, options, timeout);
       }
       throw err;
@@ -75,16 +79,31 @@ class Balancer {
 
   async getNext() {
     const proxies = await this.getProxies();
-    if(proxies.length === 0) {
+    if (proxies.length === 0) {
       throw new Error("Empty proxy list");
     }
     const proxy = proxies[this.next];
-    if(proxy) {
+    if (proxy) {
       this.next = this.next + 1;
       return proxy;
     }
     this.next = 0;
     return proxies[0];
+  }
+
+  async fetch(url, options, timeout = 5) {
+    try {
+      const res = await this.requestor(url, {
+        ...options,
+        timeout: timeout * 1000
+      })
+      if (!res.ok) {
+        throw new Error('Invalid Response');
+      }
+      return res;
+    } catch (err) {
+      throw err;
+    }
   }
 }
