--- lib/utils.js
@@ -1,7 +1,0 @@
-function delay(ms) {
-  return new Promise(resolve => setTimeout(resolve, ms));
-}
-
-module.exports = {
-  delay
-}
